// FE/src/app/model/datamodel.model.ts
export interface DataModel {
    calc_field1: number;
    calc_field1_transformation: string;
    calc_field2: number;
    mid_operator: string;
    calc_field2_transformation: string;
    enable_post_operator: boolean;
    post_operator: string;
    post_operator_val: number;
    field_id: number;
  }
  